refactor(registry): split report analytics (v1) under 200-line policy
